Rename 'nft/collection' package to 'nft/collections'
